+++ PHPCompatibility/Helpers/ResolveHelper.php
@@ -13,6 +13,7 @@ namespace PHPCompatibility\Helpers;
 use PHP_CodeSniffer\Files\File;
 use PHP_CodeSniffer\Util\Tokens;
 use PHPCSUtils\Utils\Namespaces;
+use PHPCSUtils\Utils\ObjectDeclarations;
 
 /**
  * Helper functions to figure out the fully qualified name/namespace of a token.
@@ -83,6 +84,44 @@ final class ResolveHelper
         $className = \trim($className);
 
         return self::getFQName($phpcsFile, $stackPtr, $className);
+    }
+
+    /**
+     * Returns the fully qualified name of the class that the specified class extends.
+     *
+     * Returns an empty string if the class does not extend another class or if
+     * the class name could not be reliably inferred.
+     *
+     * @since 7.0.3
+     * @since 10.0.0 This method is now static.
+     *
+     * @param \PHP_CodeSniffer\Files\File $phpcsFile The file being scanned.
+     * @param int                         $stackPtr  The position of a T_CLASS token.
+     *
+     * @return string
+     */
+    public static function getFQExtendedClassName(File $phpcsFile, $stackPtr)
+    {
+        $tokens = $phpcsFile->getTokens();
+
+        // Check for the existence of the token.
+        if (isset($tokens[$stackPtr]) === false) {
+            return '';
+        }
+
+        if ($tokens[$stackPtr]['code'] !== \T_CLASS
+            && $tokens[$stackPtr]['code'] !== \T_ANON_CLASS
+            && $tokens[$stackPtr]['code'] !== \T_INTERFACE
+        ) {
+            return '';
+        }
+
+        $extends = ObjectDeclarations::findExtendedClassName($phpcsFile, $stackPtr);
+        if (empty($extends) || \is_string($extends) === false) {
+            return '';
+        }
+
+        return self::getFQName($phpcsFile, $stackPtr, $extends);
     }
 
     /**
