+++ PHPCompatibility/Util/Tests/Helpers/ResolveHelper/GetFQExtendedClassNameUnitTest.inc
